Simplify namespace traversal and flattening
The previous approach required that when selecting an element the corresponding
NSContext be captured as well for future use.

Instead, we now have NSSelect* methods which will re-declare namespaces on any
selected element.

This also introduces a more generic internal traversal mechanism which will
walk an etree, and trigger a callback for each element with its surrounding
context.